feat: add solutions to lc problem: No.1878
No.1878.Get Biggest Three Rhombus Sums in a Grid
