+++ solution/1800-1899/1878.Get Biggest Three Rhombus Sums in a Grid/Solution.ts
@@ -1,249 +1,3 @@
-# [1878. 矩阵中最大的三个菱形和](https://leetcode.cn/problems/get-biggest-three-rhombus-sums-in-a-grid)
-
-[English Version](/solution/1800-1899/1878.Get%20Biggest%20Three%20Rhombus%20Sums%20in%20a%20Grid/README_EN.md)
-
-## 题目描述
-
-<!-- 这里写题目描述 -->
-
-<p>给你一个 <code>m x n</code> 的整数矩阵 <code>grid</code> 。</p>
-
-<p><strong>菱形和</strong> 指的是 <code>grid</code> 中一个正菱形 <strong>边界</strong> 上的元素之和。本题中的菱形必须为正方形旋转45度，且四个角都在一个格子当中。下图是四个可行的菱形，每个菱形和应该包含的格子都用了相应颜色标注在图中。</p>
-<img alt="" src="https://fastly.jsdelivr.net/gh/doocs/leetcode@main/solution/1800-1899/1878.Get%20Biggest%20Three%20Rhombus%20Sums%20in%20a%20Grid/images/pc73-q4-desc-2.png" style="width: 385px; height: 385px;" />
-<p> </p>
-
-<p>注意，菱形可以是一个面积为 0 的区域，如上图中右下角的紫色菱形所示。</p>
-
-<p>请你按照 <strong>降序</strong> 返回 <code>grid</code> 中三个最大的 <strong>互不相同的菱形和</strong> 。如果不同的和少于三个，则将它们全部返回。</p>
-
-<p> </p>
-
-<p><strong>示例 1：</strong></p>
-<img alt="" src="https://fastly.jsdelivr.net/gh/doocs/leetcode@main/solution/1800-1899/1878.Get%20Biggest%20Three%20Rhombus%20Sums%20in%20a%20Grid/images/pc73-q4-ex1.png" style="width: 360px; height: 361px;" />
-<pre>
-<b>输入：</b>grid = [[3,4,5,1,3],[3,3,4,2,3],[20,30,200,40,10],[1,5,5,4,1],[4,3,2,2,5]]
-<b>输出：</b>[228,216,211]
-<b>解释：</b>最大的三个菱形和如上图所示。
-- 蓝色：20 + 3 + 200 + 5 = 228
-- 红色：200 + 2 + 10 + 4 = 216
-- 绿色：5 + 200 + 4 + 2 = 211
-</pre>
-
-<p><strong>示例 2：</strong></p>
-<img alt="" src="https://fastly.jsdelivr.net/gh/doocs/leetcode@main/solution/1800-1899/1878.Get%20Biggest%20Three%20Rhombus%20Sums%20in%20a%20Grid/images/pc73-q4-ex2.png" style="width: 217px; height: 217px;" />
-<pre>
-<b>输入：</b>grid = [[1,2,3],[4,5,6],[7,8,9]]
-<b>输出：</b>[20,9,8]
-<b>解释：</b>最大的三个菱形和如上图所示。
-- 蓝色：4 + 2 + 6 + 8 = 20
-- 红色：9 （右下角红色的面积为 0 的菱形）
-- 绿色：8 （下方中央面积为 0 的菱形）
-</pre>
-
-<p><strong>示例 3：</strong></p>
-
-<pre>
-<b>输入：</b>grid = [[7,7,7]]
-<b>输出：</b>[7]
-<b>解释：</b>所有三个可能的菱形和都相同，所以返回 [7] 。
-</pre>
-
-<p> </p>
-
-<p><strong>提示：</strong></p>
-
-<ul>
-	<li><code>m == grid.length</code></li>
-	<li><code>n == grid[i].length</code></li>
-	<li><code>1 <= m, n <= 100</code></li>
-	<li><code>1 <= grid[i][j] <= 10<sup>5</sup></code></li>
-</ul>
-
-## 解法
-
-<!-- 这里可写通用的实现逻辑 -->
-
-**方法一：枚举菱形中心点 + 前缀和 + 有序集合**
-
-我们可以预处理得到两个前缀和数组 $s_1$ 和 $s_2$，其中 $s_1[i][j]$ 表示以 $(i, j)$ 为末尾的左上对角线上的元素之和，而 $s_2[i][j]$ 表示以 $(i, j)$ 为末尾的右上对角线上的元素之和。
-
-接下来，我们枚举每个位置 $(i, j)$，先将 $grid[i][j]$ 加入有序集合 $ss$ 中，然后枚举菱形的边长 $k$，那么以 $(i, j)$ 为中心点、边长为 $k$ 的菱形的和为：
-
-$$
-\begin{aligned}
-&\quad s_1[i + k][j] - s_1[i][j - k] + s_1[i][j + k] - s_1[i - k][j] \\
-&+ s_2[i][j - k] - s_2[i - k][j] + s_2[i + k][j] - s_2[i][j + k] \\
-&- grid[i + k - 1][j - 1] + grid[i - k - 1][j - 1]
-\end{aligned}
-$$
-
-我们将这个值加入有序集合 $ss$ 中，同时保证有序集合 $ss$ 的大小不超过 $3$，最后将有序集合 $ss$ 中的元素逆序输出即可。
-
-时间复杂度 $O(m \times n \times \min(m, n))$，空间复杂度 $O(m \times n)$。其中 $m$ 和 $n$ 分别是矩阵的行数和列数。
-
-<!-- tabs:start -->
-
-### **Python3**
-
-<!-- 这里可写当前语言的特殊实现逻辑 -->
-
-```python
-from sortedcontainers import SortedSet
-
-
-class Solution:
-    def getBiggestThree(self, grid: List[List[int]]) -> List[int]:
-        m, n = len(grid), len(grid[0])
-        s1 = [[0] * (n + 2) for _ in range(m + 1)]
-        s2 = [[0] * (n + 2) for _ in range(m + 1)]
-        for i, row in enumerate(grid, 1):
-            for j, x in enumerate(row, 1):
-                s1[i][j] = s1[i - 1][j - 1] + x
-                s2[i][j] = s2[i - 1][j + 1] + x
-        ss = SortedSet()
-        for i, row in enumerate(grid, 1):
-            for j, x in enumerate(row, 1):
-                l = min(i - 1, m - i, j - 1, n - j)
-                ss.add(x)
-                for k in range(1, l + 1):
-                    a = s1[i + k][j] - s1[i][j - k]
-                    b = s1[i][j + k] - s1[i - k][j]
-                    c = s2[i][j - k] - s2[i - k][j]
-                    d = s2[i + k][j] - s2[i][j + k]
-                    ss.add(
-                        a + b + c + d - grid[i + k - 1][j - 1] + grid[i - k - 1][j - 1]
-                    )
-                while len(ss) > 3:
-                    ss.remove(ss[0])
-        return list(ss)[::-1]
-```
-
-### **Java**
-
-<!-- 这里可写当前语言的特殊实现逻辑 -->
-
-```java
-class Solution {
-    public int[] getBiggestThree(int[][] grid) {
-        int m = grid.length, n = grid[0].length;
-        int[][] s1 = new int[m + 1][n + 2];
-        int[][] s2 = new int[m + 1][n + 2];
-        for (int i = 1; i <= m; ++i) {
-            for (int j = 1; j <= n; ++j) {
-                s1[i][j] = s1[i - 1][j - 1] + grid[i - 1][j - 1];
-                s2[i][j] = s2[i - 1][j + 1] + grid[i - 1][j - 1];
-            }
-        }
-        TreeSet<Integer> ss = new TreeSet<>();
-        for (int i = 1; i <= m; ++i) {
-            for (int j = 1; j <= n; ++j) {
-                int l = Math.min(Math.min(i - 1, m - i), Math.min(j - 1, n - j));
-                ss.add(grid[i - 1][j - 1]);
-                for (int k = 1; k <= l; ++k) {
-                    int a = s1[i + k][j] - s1[i][j - k];
-                    int b = s1[i][j + k] - s1[i - k][j];
-                    int c = s2[i][j - k] - s2[i - k][j];
-                    int d = s2[i + k][j] - s2[i][j + k];
-                    ss.add(a + b + c + d - grid[i + k - 1][j - 1] + grid[i - k - 1][j - 1]);
-                }
-                while (ss.size() > 3) {
-                    ss.pollFirst();
-                }
-            }
-        }
-        int[] ans = new int[ss.size()];
-        for (int i = 0; i < ans.length; ++i) {
-            ans[i] = ss.pollLast();
-        }
-        return ans;
-    }
-}
-```
-
-### **C++**
-
-```cpp
-class Solution {
-public:
-    vector<int> getBiggestThree(vector<vector<int>>& grid) {
-        int m = grid.size(), n = grid[0].size();
-        vector<vector<int>> s1(m + 1, vector<int>(n + 2));
-        vector<vector<int>> s2(m + 1, vector<int>(n + 2));
-        for (int i = 1; i <= m; ++i) {
-            for (int j = 1; j <= n; ++j) {
-                s1[i][j] = s1[i - 1][j - 1] + grid[i - 1][j - 1];
-                s2[i][j] = s2[i - 1][j + 1] + grid[i - 1][j - 1];
-            }
-        }
-        set<int> ss;
-        for (int i = 1; i <= m; ++i) {
-            for (int j = 1; j <= n; ++j) {
-                int l = min({i - 1, m - i, j - 1, n - j});
-                ss.insert(grid[i - 1][j - 1]);
-                for (int k = 1; k <= l; ++k) {
-                    int a = s1[i + k][j] - s1[i][j - k];
-                    int b = s1[i][j + k] - s1[i - k][j];
-                    int c = s2[i][j - k] - s2[i - k][j];
-                    int d = s2[i + k][j] - s2[i][j + k];
-                    ss.insert(a + b + c + d - grid[i + k - 1][j - 1] + grid[i - k - 1][j - 1]);
-                }
-                while (ss.size() > 3) {
-                    ss.erase(ss.begin());
-                }
-            }
-        }
-        return vector<int>(ss.rbegin(), ss.rend());
-    }
-};
-```
-
-### **Go**
-
-```go
-func getBiggestThree(grid [][]int) []int {
-	m, n := len(grid), len(grid[0])
-	s1 := make([][]int, m+1)
-	s2 := make([][]int, m+1)
-	for i := range s1 {
-		s1[i] = make([]int, n+2)
-		s2[i] = make([]int, n+2)
-	}
-	for i := 1; i <= m; i++ {
-		for j := 1; j <= n; j++ {
-			s1[i][j] = s1[i-1][j-1] + grid[i-1][j-1]
-			s2[i][j] = s2[i-1][j+1] + grid[i-1][j-1]
-		}
-	}
-
-	ss := treemap.NewWithIntComparator()
-	for i := 1; i <= m; i++ {
-		for j := 1; j <= n; j++ {
-			l := min(i-1, m-i, j-1, n-j)
-			ss.Put(grid[i-1][j-1], nil)
-			for k := 1; k <= l; k++ {
-				a := s1[i+k][j] - s1[i][j-k]
-				b := s1[i][j+k] - s1[i-k][j]
-				c := s2[i][j-k] - s2[i-k][j]
-				d := s2[i+k][j] - s2[i][j+k]
-				ss.Put(a+b+c+d-grid[i+k-1][j-1]+grid[i-k-1][j-1], nil)
-			}
-			for ss.Size() > 3 {
-				x, _ := ss.Min()
-				ss.Remove(x.(int))
-			}
-		}
-	}
-	ans := make([]int, ss.Size())
-	for i, k := range ss.Keys() {
-		ans[len(ans)-i-1] = k.(int)
-	}
-	return ans
-}
-```
-
-### **TypeScript**
-
-```ts
 function getBiggestThree(grid: number[][]): number[] {
     const m = grid.length;
     const n = grid[0].length;
@@ -914,12 +668,3 @@ class TreeMultiSet<T = number> {
         return undefined;
     }
 }
-```
-
-### **...**
-
-```
-
-```
-
-<!-- tabs:end -->
